requirements.txt
# FastAPI and server
fastapi>=0.103.0
uvicorn>=0.23.2
pydantic>=2.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.24.1

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database and storage
supabase>=1.0.3
sqlalchemy>=2.0.20

# Vector database
pinecone-client>=2.2.4

# PDF processing
pymupdf>=1.23.5
langchain>=0.0.300
langchain-text-splitters>=0.0.1

# OCR
mistralai>=0.0.8
pathlib>=1.0.1
requests>=2.31.0

# LLM integration
openai>=1.3.0
anthropic>=0.5.0

# Document generation
python-pptx>=0.6.21
python-docx>=0.8.11
reportlab>=4.0.4
fpdf2>=2.7.5
pandas>=2.0.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0